remove emission source per post graph for beges export
